Refactors the geoip code packages by moving more classes into the extension package. Also make quite a few projects package protected as they are specific to the extension.
