+++ data-prepper-plugins/geoip-processor/src/main/java/org/opensearch/dataprepper/plugins/geoip/extension/api/GeoIpConfigSupplier.java
@@ -3,7 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.plugins.geoip.extension;
+package org.opensearch.dataprepper.plugins.geoip.extension.api;
+
+import org.opensearch.dataprepper.plugins.geoip.extension.GeoIPProcessorService;
+import org.opensearch.dataprepper.plugins.geoip.extension.GeoIpConfigExtension;
 
 import java.util.Optional;
