+++ data-prepper-plugins/geoip-processor/src/main/java/org/opensearch/dataprepper/plugins/geoip/processor/GeoIPProcessor.java
@@ -18,11 +18,11 @@ import org.opensearch.dataprepper.model.processor.Processor;
 import org.opensearch.dataprepper.model.record.Record;
 import org.opensearch.dataprepper.plugins.geoip.GeoIPDatabase;
 import org.opensearch.dataprepper.plugins.geoip.GeoIPField;
-import org.opensearch.dataprepper.plugins.geoip.databaseenrich.GeoIPDatabaseReader;
 import org.opensearch.dataprepper.plugins.geoip.exception.EngineFailureException;
 import org.opensearch.dataprepper.plugins.geoip.exception.EnrichFailedException;
 import org.opensearch.dataprepper.plugins.geoip.extension.GeoIPProcessorService;
-import org.opensearch.dataprepper.plugins.geoip.extension.GeoIpConfigSupplier;
+import org.opensearch.dataprepper.plugins.geoip.extension.api.GeoIPDatabaseReader;
+import org.opensearch.dataprepper.plugins.geoip.extension.api.GeoIpConfigSupplier;
 import org.opensearch.dataprepper.plugins.geoip.utils.IPValidationCheck;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
