+++ data-prepper-plugins/geoip-processor/src/test/java/org/opensearch/dataprepper/plugins/geoip/extension/GeoIPProcessorServiceTest.java
@@ -10,8 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.opensearch.dataprepper.plugins.geoip.databaseenrich.GeoIPDatabaseReader;
-import org.opensearch.dataprepper.plugins.geoip.extension.databasedownload.GeoIPDatabaseManager;
+import org.opensearch.dataprepper.plugins.geoip.extension.api.GeoIPDatabaseReader;
 
 import java.time.Duration;
 import java.time.Instant;
